Added game listing feature
Created ListGameController, GameService and GameRepository with all methods necesaries to be able to list games with filters

Changed password of connection with the database in application.properties
Commented auto-creation of tables to test the application.
Moved import.sql and schema.sql to aux folder to be able to test the application.